replace uint64_t by unsigned int in OverlapIndexerBase static definitions

+++ src/libmaus2/dazzler/align/OverlapIndexerBase.hpp
@@ -26,8 +26,8 @@ namespace libmaus2
 		{
 			struct OverlapIndexerBase
 			{
-				static uint64_t const base_level_log = 8;
-				static uint64_t const inner_level_log = 3;			
+				static unsigned int const base_level_log = 8;
+				static unsigned int const inner_level_log = 3;			
 			};
 		}
 	}
